Removed old file for mts model, renamed run_mts_easy to run_mts
